requirements.txt

pydantic
jinja2
pyyaml
